am275x: uart_bootloader: update uart_bootloader script
- Update uart_bootloader script to flash
  wkup-r5 examples to ROM successfully.

Fixes: SITSW-5704

+++ tools/boot/uart_bootloader.py
@@ -409,20 +409,21 @@ def main(argv):
                     print("Sent {} of size {} bytes in {}s.".format(orig_filename, os.path.getsize(orig_filename), timetaken))
                     print(send_status)
                     print("")
-
-            magic_word_filename = "magic_word_file.dat"
-            magic_word_file     = open(magic_word_filename,"wb")
-            magic_word_bytes    = BOOTLOADER_END_OF_FILES_TRANSFER.to_bytes(BOOTLOADER_END_OF_FILES_TRANSFER_WORD_LENGTH,"big")
-            magic_word_file.write(magic_word_bytes)
-            magic_word_file.close()
-            sendd_status, timetaken = xmodem_send_receive_file(magic_word_filename, serialport, get_response=False)
-            print("")
-            print(" Sent End Of File Transfer message of size {} bytes in {}s.".format( os.path.getsize(magic_word_filename), timetaken))
-            print("")
-            os.remove(magic_word_filename)
-            print("All commands from config file are executed !!!")
-            if("SUCCESS" in send_status):
-                print("Connect to UART in 5 seconds to see logs from UART !!!")
+            
+            if(length != 1):
+                magic_word_filename = "magic_word_file.dat"
+                magic_word_file     = open(magic_word_filename,"wb")
+                magic_word_bytes    = BOOTLOADER_END_OF_FILES_TRANSFER.to_bytes(BOOTLOADER_END_OF_FILES_TRANSFER_WORD_LENGTH,"big")
+                magic_word_file.write(magic_word_bytes)
+                magic_word_file.close()
+                sendd_status, timetaken = xmodem_send_receive_file(magic_word_filename, serialport, get_response=False)
+                print("")
+                print(" Sent End Of File Transfer message of size {} bytes in {}s.".format( os.path.getsize(magic_word_filename), timetaken))
+                print("")
+                os.remove(magic_word_filename)
+                print("All commands from config file are executed !!!")
+                if("SUCCESS" in send_status):
+                    print("Connect to UART in 5 seconds to see logs from UART !!!")
 
     else:
         # Validate the cmdline config
